requirements.txt
# Core dependencies
pyyaml>=6.0
psutil>=5.9.0

# For distributed coordination
redis>=4.5.0  # Optional: for distributed queue
requests>=2.31.0

# For monitoring and metrics
matplotlib>=3.7.0
numpy>=1.24.0
pandas>=2.0.0

# For logging
coloredlogs>=15.0

# For testing
pytest>=7.4.0
pytest-cov>=4.1.0
